+++ package.json
@@ -14,10 +14,12 @@
 		"dev": "tsc --watch",
 		"test": "jest",
 		"test:bail": "jest --bail",
-		"test:watch": "jest --watch",
 		"test:coverage": "jest --coverage",
-		"lint": "eslint src --ext .ts",
-		"lint:fix": "eslint src --ext .ts --fix",
+		"test:subduction": "jest test/unit/repo-factory.test.ts test/unit/network-sync-sub.test.ts test/unit/subduction-config.test.ts test/integration/sub-flag.test.ts",
+		"test:watch": "jest --watch",
+		"test:websocket": "jest --testPathIgnorePatterns=/node_modules/ --testPathIgnorePatterns=\"(repo-factory|network-sync-sub|subduction-config|sub-flag)[.]test[.]ts$\"",
+		"lint": "eslint src",
+		"lint:fix": "eslint src --fix",
 		"clean": "rm -rf dist",
 		"prepack": "pnpm run build",
 		"start": "node dist/cli.js",
@@ -61,12 +63,14 @@
 		"@types/node": "^20.0.0",
 		"@types/tmp": "^0.2.4",
 		"babel-jest": "^30.2.0",
+		"eslint": "^9.18.0",
 		"fast-check": "^4.3.0",
 		"jest": "^29.7.0",
 		"tmp": "^0.2.1",
 		"ts-jest": "^29.1.0",
 		"tsx": "^4.19.2",
-		"typescript": "^5.2.0"
+		"typescript": "^5.2.0",
+		"typescript-eslint": "^8.20.0"
 	},
 	"jest": {
 		"preset": "ts-jest",
@@ -86,6 +90,7 @@
 		"setupFilesAfterEnv": [
 			"<rootDir>/test/jest.setup.ts"
 		],
+		"globalSetup": "<rootDir>/test/jest.globalSetup.ts",
 		"maxWorkers": "75%",
 		"maxConcurrency": 10,
 		"transformIgnorePatterns": [
