+++ test/unit/repo-factory.test.ts
@@ -14,14 +14,12 @@ import * as fs from "fs/promises";
 import * as tmp from "tmp";
 import { execSync } from "child_process";
 
+// The CLI bundle (`dist/cli.js`) is built once by `test/jest.globalSetup.ts`
+// before any worker spawns; no per-suite build is needed here.
 describe("createRepo with --sub", () => {
   let tmpDir: string;
   let cleanup: () => void;
   const cliPath = path.join(__dirname, "../../dist/cli.js");
-
-  beforeAll(() => {
-    execSync("pnpm build", { cwd: path.join(__dirname, "../.."), stdio: "pipe" });
-  });
 
   beforeEach(async () => {
     const tmpObj = tmp.dirSync({ unsafeCleanup: true });
@@ -33,79 +31,104 @@ describe("createRepo with --sub", () => {
     cleanup();
   });
 
-  it("should create a working repo with --sub flag", async () => {
-    await fs.writeFile(path.join(tmpDir, "test.txt"), "hello");
+  // `init --sub` and `sync` against the live Subduction server can take
+  // longer than Jest's default 5s test timeout. The execSync timeout is
+  // kept slightly below the Jest per-test timeout so that a network stall
+  // surfaces as the underlying CLI error rather than Jest's generic
+  // "exceeded timeout" message.
+  const INIT_TIMEOUT_MS = 50000;
+  const FAST_TIMEOUT_MS = 15000;
+  const JEST_TIMEOUT_MS = 60000;
 
-    execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
-      stdio: "pipe",
-      timeout: 30000,
-    });
+  it(
+    "should create a working repo with --sub flag",
+    async () => {
+      await fs.writeFile(path.join(tmpDir, "test.txt"), "hello");
 
-    const snapshotPath = path.join(tmpDir, ".pushwork", "snapshot.json");
-    const stat = await fs.stat(snapshotPath);
-    expect(stat.isFile()).toBe(true);
-  });
+      execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
+        stdio: "pipe",
+        timeout: INIT_TIMEOUT_MS,
+      });
 
-  it("should produce a valid automerge URL", async () => {
-    await fs.writeFile(path.join(tmpDir, "test.txt"), "hello");
+      const snapshotPath = path.join(tmpDir, ".pushwork", "snapshot.json");
+      const stat = await fs.stat(snapshotPath);
+      expect(stat.isFile()).toBe(true);
+    },
+    JEST_TIMEOUT_MS
+  );
 
-    execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
-      stdio: "pipe",
-      timeout: 30000,
-    });
+  it(
+    "should produce a valid automerge URL",
+    async () => {
+      await fs.writeFile(path.join(tmpDir, "test.txt"), "hello");
 
-    const url = execSync(`node "${cliPath}" url "${tmpDir}"`, {
-      encoding: "utf8",
-      timeout: 10000,
-    }).trim();
+      execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
+        stdio: "pipe",
+        timeout: INIT_TIMEOUT_MS,
+      });
 
-    expect(url).toMatch(/^automerge:/);
-  });
+      const url = execSync(`node "${cliPath}" url "${tmpDir}"`, {
+        encoding: "utf8",
+        timeout: FAST_TIMEOUT_MS,
+      }).trim();
 
-  it("should track files in the snapshot", async () => {
-    await fs.writeFile(path.join(tmpDir, "a.txt"), "aaa");
-    await fs.mkdir(path.join(tmpDir, "sub"), { recursive: true });
-    await fs.writeFile(path.join(tmpDir, "sub", "b.txt"), "bbb");
+      expect(url).toMatch(/^automerge:/);
+    },
+    JEST_TIMEOUT_MS
+  );
 
-    execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
-      stdio: "pipe",
-      timeout: 30000,
-    });
+  it(
+    "should track files in the snapshot",
+    async () => {
+      await fs.writeFile(path.join(tmpDir, "a.txt"), "aaa");
+      await fs.mkdir(path.join(tmpDir, "sub"), { recursive: true });
+      await fs.writeFile(path.join(tmpDir, "sub", "b.txt"), "bbb");
 
-    const ls = execSync(`node "${cliPath}" ls "${tmpDir}"`, {
-      encoding: "utf8",
-      timeout: 10000,
-    });
+      execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
+        stdio: "pipe",
+        timeout: INIT_TIMEOUT_MS,
+      });
 
-    expect(ls).toContain("a.txt");
-    expect(ls).toContain("b.txt");
-  });
+      const ls = execSync(`node "${cliPath}" ls "${tmpDir}"`, {
+        encoding: "utf8",
+        timeout: FAST_TIMEOUT_MS,
+      });
 
-  it("should be able to sync after init", async () => {
-    await fs.writeFile(path.join(tmpDir, "initial.txt"), "first");
+      expect(ls).toContain("a.txt");
+      expect(ls).toContain("b.txt");
+    },
+    JEST_TIMEOUT_MS
+  );
 
-    execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
-      stdio: "pipe",
-      timeout: 30000,
-    });
+  it(
+    "should be able to sync after init",
+    async () => {
+      await fs.writeFile(path.join(tmpDir, "initial.txt"), "first");
 
-    // Add a new file
-    await fs.writeFile(path.join(tmpDir, "added.txt"), "second");
+      execSync(`node "${cliPath}" init --sub "${tmpDir}"`, {
+        stdio: "pipe",
+        timeout: INIT_TIMEOUT_MS,
+      });
 
-    // Sync should not throw. The `sync` command has no --sub flag — it
-    // reads the backend choice from .pushwork/config.json (persisted by
-    // the init --sub above).
-    execSync(`node "${cliPath}" sync "${tmpDir}"`, {
-      stdio: "pipe",
-      timeout: 30000,
-    });
+      // Add a new file
+      await fs.writeFile(path.join(tmpDir, "added.txt"), "second");
 
-    const ls = execSync(`node "${cliPath}" ls "${tmpDir}"`, {
-      encoding: "utf8",
-      timeout: 10000,
-    });
+      // Sync should not throw. The `sync` command has no --sub flag — it
+      // reads the backend choice from .pushwork/config.json (persisted by
+      // the init --sub above).
+      execSync(`node "${cliPath}" sync "${tmpDir}"`, {
+        stdio: "pipe",
+        timeout: INIT_TIMEOUT_MS,
+      });
 
-    expect(ls).toContain("initial.txt");
-    expect(ls).toContain("added.txt");
-  });
+      const ls = execSync(`node "${cliPath}" ls "${tmpDir}"`, {
+        encoding: "utf8",
+        timeout: FAST_TIMEOUT_MS,
+      });
+
+      expect(ls).toContain("initial.txt");
+      expect(ls).toContain("added.txt");
+    },
+    JEST_TIMEOUT_MS
+  );
 });
